+++ client/src/main/java/client/scenes/AddActivityCtrl.java
@@ -79,7 +79,7 @@ public class AddActivityCtrl {
             var imagePathText = image_path.getText();
             var titleText = title.getText();
             if (titleText.length() > 140) return null;
-            var consumptionInt = Integer.parseInt(consumption.getText());
+            var consumptionInt = Long.parseLong(consumption.getText());
             if (consumptionInt == 0) return null;
             var sourceText = source.getText();
             return new Activity(idText, imagePathText, titleText, consumptionInt, sourceText);
